refactor the CLI logic interface to support both package layouts
