+++ README.md
@@ -81,7 +81,7 @@ add the new ones in the current directory:
 
 ```sh
 bower cache clean && bower install
-modulizer --out .
+modulizer package --out .
 ```
 
 ### Workspace mode
@@ -91,7 +91,7 @@ You must first generate a GitHub access token and store it in a file named `gith
 Then run:
 
 ```sh
-modulizer owner/repo owner2/repo2
+modulizer workspace --repo owner/repo owner2/repo2
 ```
 
 This will create a `modulizer_workspace` directory and checkout the repos and their Bower dependencies and convert them all in place. You can then run `polymer serve` in the workspace directory and try out the results in Chrome 61 or Safari 10.1 (or Edge and Firefox with the appropriate flags turned on).
